fix: changes contact-us templates structure to be consistent with homepage templates structure
